+++ src/utils/file/config_reader.rs
@@ -2,7 +2,10 @@ use crate::m3u_filter_error::{create_m3u_filter_error, create_m3u_filter_error_r
 use crate::model::api_proxy::ApiProxyConfig;
 use crate::model::config::{Config, ConfigDto, ConfigInputAlias, InputType};
 use crate::model::mapping::Mappings;
+use crate::utils::file::env_resolving_reader::EnvResolvingReader;
+use crate::utils::file::file_utils::file_reader;
 use crate::utils::file::{file_utils, multi_file_reader};
+use crate::utils::network::request::get_credentials_from_url;
 use crate::utils::sys_utils::exit;
 use chrono::Local;
 use log::{debug, error, info, warn};
@@ -10,16 +13,39 @@ use regex::Regex;
 use serde::Serialize;
 use std::env;
 use std::fs::File;
-use std::io::{self, BufRead};
+use std::io::{self, BufRead, BufReader, Read};
 use std::path::PathBuf;
 use std::sync::LazyLock;
-use url::{Url};
-use crate::utils::network::request::get_credentials_from_url;
+use url::Url;
 
-pub fn read_mappings(args_mapping: Option<String>, cfg: &mut Config) -> Result<Option<String>, M3uFilterError> {
+enum EitherReader<L, R> {
+    Left(L),
+    Right(R),
+}
+
+// `Read`-Trait für Either implementieren
+impl<L: Read, R: Read> Read for EitherReader<L, R> {
+    fn read(&mut self, buf: &mut [u8]) -> io::Result<usize> {
+        match self {
+            EitherReader::Left(reader) => reader.read(buf),
+            EitherReader::Right(reader) => reader.read(buf),
+        }
+    }
+}
+
+pub fn config_file_reader(file: File, resolve_env: bool) -> impl Read
+{
+    if resolve_env {
+        EitherReader::Left(EnvResolvingReader::new(file_reader(file)))
+    } else {
+        EitherReader::Right(BufReader::new(file))
+    }
+}
+
+pub fn read_mappings(args_mapping: Option<String>, cfg: &mut Config, resolve_env: bool) -> Result<Option<String>, M3uFilterError> {
     let mappings_file: String = args_mapping.unwrap_or_else(|| file_utils::get_default_mappings_path(cfg.t_config_path.as_str()));
 
-    match read_mapping(mappings_file.as_str()) {
+    match read_mapping(mappings_file.as_str(), resolve_env) {
         Ok(mappings) => {
             match mappings {
                 None => {
@@ -55,13 +81,13 @@ pub async fn read_api_proxy_config(args_api_proxy_config: Option<String>, cfg: &
 pub fn read_config(config_path: &str, config_file: &str, sources_file: &str) -> Result<Config, M3uFilterError> {
     let files = vec![std::path::PathBuf::from(config_file), std::path::PathBuf::from(sources_file)];
     match multi_file_reader::MultiFileReader::new(&files) {
-        Ok(file) => {
-            match serde_yaml::from_reader::<_, Config>(file) {
+        Ok(reader) => {
+            match serde_yaml::from_reader::<_, Config>(reader) {
                 Ok(mut result) => {
                     result.t_config_path = config_path.to_string();
                     result.t_config_file_path = config_file.to_string();
                     result.t_sources_file_path = sources_file.to_string();
-                    match result.prepare(true) {
+                    match result.prepare() {
                         Err(err) => Err(err),
                         _ => Ok(result),
                     }
@@ -75,10 +101,10 @@ pub fn read_config(config_path: &str, config_file: &str, sources_file: &str) -> 
     }
 }
 
-pub fn read_mapping(mapping_file: &str) -> Result<Option<Mappings>, M3uFilterError> {
+pub fn read_mapping(mapping_file: &str, resolve_var: bool) -> Result<Option<Mappings>, M3uFilterError> {
     let mapping_file = std::path::PathBuf::from(mapping_file);
     if let Ok(file) = file_utils::open_file(&mapping_file) {
-        let mapping: Result<Mappings, _> = serde_yaml::from_reader(file);
+        let mapping: Result<Mappings, _> = serde_yaml::from_reader(config_file_reader(file, resolve_var));
         match mapping {
             Ok(mut result) => {
                 handle_m3u_filter_error_result!(M3uFilterErrorKind::Info, result.prepare());
@@ -93,12 +119,12 @@ pub fn read_mapping(mapping_file: &str) -> Result<Option<Mappings>, M3uFilterErr
     Ok(None)
 }
 
-pub fn read_api_proxy(config: &Config, api_proxy_file: &str, resolve_var: bool) -> Option<ApiProxyConfig> {
+pub fn read_api_proxy(config: &Config, api_proxy_file: &str, resolve_env: bool) -> Option<ApiProxyConfig> {
     file_utils::open_file(&std::path::PathBuf::from(api_proxy_file)).map_or(None, |file| {
-        let mapping: Result<ApiProxyConfig, _> = serde_yaml::from_reader(file);
+        let mapping: Result<ApiProxyConfig, _> = serde_yaml::from_reader(config_file_reader(file, resolve_env));
         match mapping {
             Ok(mut result) => {
-                match result.prepare(config, resolve_var) {
+                match result.prepare(config) {
                     Err(err) => {
                         exit!("cant read api-proxy-config file: {}", err);
                     }
@@ -155,16 +181,6 @@ pub fn resolve_env_var(value: &str) -> String {
     }).to_string()
 }
 
-//
-// pub fn read_yaml_inputs(input_type: InputType, file_path:  &str) -> Result<Config, M3uFilterError> {
-//
-// }
-//
-//
-// pub fn read_json_inputs(input_type: InputType, file_path: &str) -> Result<Config, M3uFilterError> {
-//
-// }
-
 const CSV_SEPARATOR: char = ';';
 const HEADER_PREFIX: char = '#';
 const FIELD_MAX_CON: &str = "max_connections";
@@ -174,23 +190,37 @@ const FIELD_NAME: &str = "name";
 const FIELD_USERNAME: &str = "username";
 const FIELD_PASSWORD: &str = "password";
 const FIELD_UNKNOWN: &str = "?";
+const DEFAULT_COLUMNS: &[&str] = &[FIELD_URL, FIELD_MAX_CON, FIELD_PRIO, FIELD_NAME, FIELD_USERNAME, FIELD_PASSWORD];
 
-fn csv_assign_mandatory_fields(alias: &mut ConfigInputAlias) {
-    if alias.name.is_empty() && !alias.url.is_empty() {
-        match  Url::parse(alias.url.as_str()) {
+fn csv_assign_mandatory_fields(alias: &mut ConfigInputAlias, input_type: &InputType) {
+    if !alias.url.is_empty() {
+        match Url::parse(alias.url.as_str()) {
             Ok(url) => {
                 let (username, password) = get_credentials_from_url(&url);
                 if username.is_none() || password.is_none() {
                     // xtream url
-                    alias.url = url.origin().ascii_serialization().to_string();
+                    if input_type == &InputType::XtreamBatch {
+                        alias.url = url.origin().ascii_serialization().to_string();
+                    } else if input_type == &InputType::M3uBatch {
+                        if alias.username.is_some() && alias.password.is_some() {
+                            alias.url = format!("{}/get_php?username={}&password={}&type=m3u_plus",
+                                                url.origin().ascii_serialization().to_string(),
+                                                alias.username.as_deref().unwrap_or("").to_string(),
+                                                alias.password.as_deref().unwrap_or("").to_string()
+                            )
+                        }
+                    }
                 } else {
+                    if input_type == &InputType::XtreamBatch {
+                        alias.url = url.origin().ascii_serialization().to_string();
+                    }
                     // m3u url
                     alias.username = username;
                     alias.password = password;
                 }
 
                 if alias.name.is_empty() {
-                    let username = alias.username.as_ref().map(|s|s.as_str()).unwrap_or_default();
+                    let username = alias.username.as_ref().map(|s| s.as_str()).unwrap_or_default();
                     let domain: Vec<&str> = url.domain().unwrap_or_default().split('.').collect();
                     if domain.len() > 1 {
                         alias.name = format!("{}_{username}", domain[domain.len() - 2]);
@@ -204,96 +234,166 @@ fn csv_assign_mandatory_fields(alias: &mut ConfigInputAlias) {
     }
 }
 
-fn csv_assign_config_input_column(config_input: &mut ConfigInputAlias, input_type: &InputType, header: &str, value: &str) -> Result<(), io::Error> {
+fn csv_assign_config_input_column(config_input: &mut ConfigInputAlias, header: &str, raw_value: &str) -> Result<(), io::Error> {
+    let value = raw_value.trim();
+    if !value.is_empty() {
         match header {
             FIELD_URL => {
                 let url = Url::parse(value.trim()).map_err(to_io_error)?;
-                match input_type {
-                    InputType::Xtream | InputType::M3u => {
-                        config_input.url = url.to_string();
-                    }
-                    InputType::M3uBatch | InputType::XtreamBatch => {
-                    }
-                };
-            },
+                config_input.url = url.to_string();
+            }
             FIELD_MAX_CON => {
                 let max_connections = value.parse::<u16>().unwrap_or(1);
                 config_input.max_connections = max_connections;
-            },
+            }
             FIELD_PRIO => {
                 let priority = value.parse::<i16>().unwrap_or(0);
                 config_input.priority = priority;
-            },
+            }
             FIELD_NAME => {
                 config_input.name = value.to_string();
-            },
+            }
+            FIELD_USERNAME => {
+                config_input.username = Some(value.to_string());
+            }
+            FIELD_PASSWORD => {
+                config_input.password = Some(value.to_string());
+            }
             _ => {}
         }
+    }
     Ok(())
 }
 
-pub fn csv_read_inputs(input_type: InputType, file_uri: &str) -> Result<Vec<ConfigInputAlias>, io::Error> {
+pub fn csv_read_inputs_from_reader(batch_input_type: InputType, reader: impl BufRead) -> Result<Vec<ConfigInputAlias>, io::Error> {
+    let input_type = match batch_input_type {
+        InputType::M3uBatch  | InputType::M3u => InputType::M3uBatch,
+        InputType::XtreamBatch | InputType::Xtream => InputType::XtreamBatch
+    };
+    let mut result = vec![];
+    let mut default_columns = vec![];
+    default_columns.extend_from_slice(DEFAULT_COLUMNS);
+    for line in reader.lines().into_iter() {
+        let line = line?;
+        if line.is_empty() {
+            continue
+        }
+        if line.starts_with(HEADER_PREFIX) {
+            default_columns = line[1..].split(CSV_SEPARATOR).map(|s| {
+                match s {
+                    FIELD_URL => FIELD_URL,
+                    FIELD_MAX_CON => FIELD_MAX_CON,
+                    FIELD_PRIO => FIELD_PRIO,
+                    FIELD_NAME => FIELD_NAME,
+                    FIELD_USERNAME => FIELD_USERNAME,
+                    FIELD_PASSWORD => FIELD_PASSWORD,
+                    _ => {
+                        error!("Field {s} is unsupported for csv input");
+                        FIELD_UNKNOWN
+                    }
+                }
+            }).collect();
+            continue;
+        }
+
+        let mut config_input = ConfigInputAlias {
+            id: 0,
+            name: "".to_string(),
+            url: "".to_string(),
+            username: None,
+            password: None,
+            priority: 0,
+            max_connections: 1,
+        };
+
+        let columns: Vec<&str> = line.split(CSV_SEPARATOR).collect();
+        for (&header, &value) in default_columns.iter().zip(columns.iter()).into_iter() {
+            if let Err(err) = csv_assign_config_input_column(&mut config_input, header, value) {
+                error!("Could not parse input line: {} err: {err}", line);
+                continue;
+            }
+        }
+        csv_assign_mandatory_fields(&mut config_input, &input_type);
+        result.push(config_input);
+    }
+    Ok(result)
+}
+
+
+pub fn csv_read_inputs(batch_input_type: InputType, file_uri: &str) -> Result<Vec<ConfigInputAlias>, io::Error> {
     let file_path = file_uri.strip_prefix("file://").unwrap_or(file_uri);
     let inputs_file = std::path::PathBuf::from(file_path);
     if let Ok(file) = file_utils::open_file(&inputs_file) {
-        let mut result = vec![];
-        let mut default_columns = vec![FIELD_URL, FIELD_MAX_CON, FIELD_PRIO, FIELD_NAME];
-        let reader = io::BufReader::new(file);
-        for line in reader.lines().into_iter() {
-            let line = line?;
-            if line.is_empty() || line.starts_with(HEADER_PREFIX) {
-                default_columns = line[1..].split(CSV_SEPARATOR).map(|s| {
-                    println!("{s}");
-                    match s {
-                        FIELD_URL => FIELD_URL,
-                        FIELD_MAX_CON => FIELD_MAX_CON,
-                        FIELD_PRIO => FIELD_PRIO,
-                        FIELD_NAME => FIELD_NAME,
-                        _ => {
-                            error!("Field {s} is unsupported for csv input");
-                            FIELD_UNKNOWN
-                        }
-                    }
-                    }).collect();
-                println!("{default_columns:?}");
-                continue;
-            }
-
-            let mut config_input = ConfigInputAlias {
-                id: 0,
-                name: "".to_string(),
-                url: "".to_string(),
-                username: None,
-                password: None,
-                priority: 0,
-                max_connections: 1,
-            };
-
-            let columns: Vec<&str> = line.split(CSV_SEPARATOR).collect();
-            for (&header, &value) in default_columns.iter().zip(columns.iter()).into_iter() {
-                if let Err(err) = csv_assign_config_input_column(&mut config_input, &input_type, header, value) {
-                    error!("Could not parse input line: {} err: {err}", line);
-                    continue;
-                }
-            }
-            csv_assign_mandatory_fields(&mut config_input);
-            result.push(config_input);
-        }
-        return Ok(result);
+        return csv_read_inputs_from_reader(batch_input_type, EnvResolvingReader::new(file_reader(file)));
     }
     Err(str_to_io_error(&format!("Could not open {file_path}")))
 }
 
 #[cfg(test)]
 mod tests {
+    use std::io::{BufReader, Cursor};
     use crate::model::config::InputType;
-    use crate::utils::file::config_reader::{csv_read_inputs, resolve_env_var};
+    use crate::utils::file::config_reader::{csv_read_inputs_from_reader, resolve_env_var};
+const M3U_BATCH: &str = r#"
+#url;name;max_connections;priority
+http://hd.providerline.com:8080/get.php?username=user1&password=user1&type=m3u_plus;input_1
+http://hd.providerline.com/get.php?username=user2&password=user2&type=m3u_plus;input_2;1;2
+http://hd.providerline.com/get.php?username=user3&password=user3&type=m3u_plus;input_3;1;2
+http://hd.providerline.com/get.php?username=user4&password=user4&type=m3u_plus;input_4
+"#;
+
+const XTREAM_BATCH: &str = r#"
+#name;username;password;url;max_connections
+input_1;desanocra;eyCG8SN523KQ;http://provider_1.tv:80;1
+input_2;desanocra;eyCG8SN523KQ;http://provider_2.tv:8080;1
+"#;
 
     #[test]
-    fn test_read_inputs() {
-        let result = csv_read_inputs(InputType::Xtream, &resolve_env_var("${env:HOME}/projects/m3u-test/settings/inputs.txt"));
+    fn test_read_inputs_xtream_as_m3u() {
+        let reader = BufReader::new(Cursor::new(XTREAM_BATCH));
+        let result = csv_read_inputs_from_reader(InputType::M3uBatch, reader);
         assert_eq!(result.is_ok(), true);
-        println!("result: {:?}", result.unwrap());
+        let aliases = result.unwrap();
+        assert_eq!(aliases.is_empty(), false);
+        for config in  aliases {
+            assert_eq!(config.url.contains("username"), true);
+        }
+    }
+
+    #[test]
+    fn test_read_inputs_m3u_as_m3u() {
+        let reader = BufReader::new(Cursor::new(M3U_BATCH));
+        let result = csv_read_inputs_from_reader(InputType::M3uBatch, reader);
+        assert_eq!(result.is_ok(), true);
+        let aliases = result.unwrap();
+        assert_eq!(aliases.is_empty(), false);
+        for config in  aliases {
+            assert_eq!(config.url.contains("username"), true);
+        }
+    }
+
+    #[test]
+    fn test_read_inputs_xtream_as_xtream() {
+        let reader = BufReader::new(Cursor::new(XTREAM_BATCH));
+        let result = csv_read_inputs_from_reader(InputType::XtreamBatch, reader);
+        assert_eq!(result.is_ok(), true);
+        let aliases = result.unwrap();
+        assert_eq!(aliases.is_empty(), false);
+        for config in  aliases {
+            assert_eq!(config.url.contains("username"), false);
+        }
+    }
+
+    #[test]
+    fn test_read_inputs_m3u_as_xtream() {
+        let reader = BufReader::new(Cursor::new(M3U_BATCH));
+        let result = csv_read_inputs_from_reader(InputType::XtreamBatch, reader);
+        assert_eq!(result.is_ok(), true);
+        let aliases = result.unwrap();
+        assert_eq!(aliases.is_empty(), false);
+        for config in  aliases {
+            assert_eq!(config.url.contains("username"), false);
+        }
     }
 
     #[test]