+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 # 2.2.3 (2023-04-xx)
+- variable resolving for config files now for all settings
 - hls reverse proxy implemented
 - dash redirect implemented (reverse proxy not supported) 
 - !BREAKING CHANGE! `channel_unavailable_file` is now under `custom_stream_response`,
@@ -27,6 +28,43 @@ custom_stream_response:
   targets:
   - name: test
 ```
+Input aliases can be defined as batches in csv files with `;` separator.
+There are 2 batch input types  `xtream_batch` and `m3u_batch`.
+`XtreamBatch`:
+
+```yaml
+- sources:
+- inputs:
+  - type: xtream_batch
+    url: 'file:///home/m3u-filter/config/my_provider_batch.csv'
+  targets:
+  - name: test
+```
+
+```csv
+#name;username;password;url;max_connections;priority
+my_provider_1;user1;password1;http://my_provider_1.com:80;1;0
+my_provider_2;user2;password2;http://my_provider_2.com:8080;1;0
+```
+
+`M3uBatch`:
+```yaml
+- sources:
+- inputs:
+  - type: m3u_batch
+    url: 'file:///home/m3u-filter/config/my_provider_batch.csv'
+    targets:
+  - name: test
+```
+
+```csv
+#name;url;max_connections;priority
+http://my_provider_1.com:80/get_php?username=user1&password=password1;1;0
+http://my_provider_2.com:8080/get_php?username=user2&password=password2;1;0
+```
+The Fields `max_connections` and `priority`are optional.
+`max_connections`  will be set default to `1`. This is different from yaml config where the default is `0=unlimited`  
+
 - added two options to reverse proxy config `forced_retry_interval_secs` and `connect_timeout_secs`
 `forced_retry_interval_secs` forces every x seconds a reconnect to the provider,
 `connect_timeout_secs` tries only x seconds for connection, if not successfully starts a retry.